Move input-related files to rsc.inputs
