+++ rsc/src/main/scala/rsc/inputs/Offsets.scala
@@ -1,6 +1,6 @@
 // Copyright (c) 2017-2018 Twitter, Inc.
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
-package rsc.lexis
+package rsc.inputs
 
 trait Offsets {
   type Offset = Int
