+++ rsc/src/main/scala/rsc/inputs/package.scala
@@ -2,4 +2,4 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc
 
-package object lexis extends Keywords with Names with Tokens
+package object inputs extends Offsets
