+++ rsc/src/main/scala/rsc/outline/Synthetics.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.outline
 
-import rsc.lexis._
+import rsc.inputs._
 import rsc.syntax._
 
 trait Synthetics {
