+++ rsc/src/main/scala/rsc/parse/Tpts.scala
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.parse
 
+import rsc.inputs._
 import rsc.lexis._
 import rsc.report._
 import rsc.syntax._
