+++ rsc/src/main/scala/rsc/pretty/PrettyInput.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.pretty
 
-import rsc.lexis._
+import rsc.inputs._
 
 object PrettyInput {
   def str(p: Printer, x: Input): Unit = {
