+++ rsc/src/main/scala/rsc/pretty/PrettyMessage.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.pretty
 
-import rsc.lexis._
+import rsc.inputs._
 import rsc.report._
 
 object PrettyMessage {
