+++ rsc/src/main/scala/rsc/pretty/PrettyPosition.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.pretty
 
-import rsc.lexis._
+import rsc.inputs._
 
 object PrettyPosition {
   def str(p: Printer, x: Position): Unit = {
