+++ rsc/src/main/scala/rsc/report/Messages.scala
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.report
 
+import rsc.inputs._
 import rsc.lexis._
 import rsc.outline._
 import rsc.pretty._
