+++ rsc/src/main/scala/rsc/report/Points.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.report
 
-import rsc.lexis._
+import rsc.inputs._
 import rsc.syntax._
 
 trait Points {
