+++ rsc/src/main/scala/rsc/scan/Messages.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.scan
 
-import rsc.lexis._
+import rsc.inputs._
 import rsc.report._
 import rsc.util._
