+++ rsc/src/main/scala/rsc/semanticdb/Semanticdb.scala
@@ -6,7 +6,7 @@ import java.io._
 import java.nio.file._
 import java.util.HashMap
 import rsc.gensym._
-import rsc.lexis._
+import rsc.inputs._
 import rsc.outline._
 import rsc.report._
 import rsc.semantics._
