+++ scalafix/rules/src/main/scala/rsc/rules/pretty/SemanticdbPrinter.scala
@@ -3,7 +3,7 @@
 // NOTE: This file has been partially copy/pasted from scalameta/scalameta.
 package rsc.rules.pretty
 
-import rsc.lexis._
+import rsc.inputs._
 import rsc.pretty._
 import rsc.rules.semantics._
 import scala.collection.mutable
